test(govern-lifecycle): SMOKE four-tier ratchet bites on growth over real src/atdd scan + audit row (#958)
