+++ docs/smoke-audit.md
@@ -230,6 +230,7 @@ of bypass patterns.
 | acc:govern-lifecycle:Y005-SMOKE-001-reconcile-wired-in-cli | real (atdd issue reconcile) | CLI wiring | N/A (single component) | — |
 | acc:govern-lifecycle:E045-SMOKE-001-real-subprocess-check-blocks-through-live-registry | real (real CommandGateCheck shells out to a real subprocess; live GateRegistry + evaluate_transition_gate; no mocks) | a genuinely-failing subprocess yields proceed=False (blocks), a genuinely-passing one yields proceed=True | CommandGateCheck → GateRegistry → evaluate_transition_gate → OS subprocess | #1020 (the enforcing-gate keystone; production wiring of the gate decision against a real process) |
 | acc:govern-lifecycle:E046-SMOKE-001-real-missing-tool-fails-closed | real (real CommandGateCheck names a binary absent from the host PATH; the OS raises FileNotFoundError) | a real missing tool is converted to a FAIL → proceed=False (fail-closed) | CommandGateCheck → GateRegistry → evaluate_transition_gate → OS exec attempt | #1020 (fail-closed guarantee verified against the real OS, not a stubbed exception) |
+| acc:govern-lifecycle:E049-SMOKE-001-new-violation-beyond-baseline-fails-the-gate | real (collect_toolkit_violations runs the production config-driven composition analysis over the real src/atdd tree; real grandfathered baseline resolved from .atdd/baselines/) | a new violation absent from the baseline leaks (fails the gate) while pre-existing grandfathered debt stays absorbed | .atdd/config.yaml code.toolkit → resolve_scan_roots → analyze_python_root (real src/atdd scan) → new_violations vs .atdd/baselines/four_tier_toolkit.yaml | #958 (the four-tier ratchet bites on growth, not on legacy; real scan, no mocks) |
 | acc:implement-code:D003-UNIT-002-no-false-positives | real (atdd validate) | no false violations | N/A (validator) | — |
 | acc:integrate-end-to-end:E001-SMOKE-001-cycle-reaches-complete | real (full ATDD cycle) | issue status | both ends (INIT → COMPLETE) | — |
 | acc:integrate-end-to-end:E001-SMOKE-002-artifacts-readable | real (full ATDD cycle) | artifact readability | both ends (write + read) | — |
